refactor (LottoService.java) : 비즈니스 로직을 service로 위임
